+++ Manager/EnemyFactory.cpp
@@ -66,7 +66,7 @@ void EnemyFactory::LoadFromCSV(StageType stageType, BulletManager* bulletManager
             int id = std::stoi(cell);
             if (id != 0 && enemyFactory.count(id)) 
             {
-                Vector2 pos{ col * cellSize, row * cellSize+kEnemyOffsetY };
+                Vector2 pos{ static_cast<float>(col * cellSize), static_cast<float>(row * cellSize+kEnemyOffsetY) };
                 enemies_.push_back(enemyFactory[id](pos));
             }
             col++;
